requirements.txt
flask>=2.3.3
flask-login>=0.6.2
flask-sqlalchemy>=3.0.5
flask-migrate>=4.0.4
flask-wtf>=1.1.1
scikit-learn>=1.3.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.2
plotly>=5.18.0
dash>=2.11.1
category_encoders>=2.6.1
lightgbm>=4.0.0
xgboost>=1.7.6
catboost>=1.2
optuna>=3.2.0
shap>=0.42.1
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
fastapi>=0.110.0
uvicorn>=0.27.0
redis>=5.0.0
celery>=5.3.1
pydantic>=2.10.2
pydantic-settings>=2.0.0
streamlit>=1.32.0
altair>=5.1.2
watchdog==3.0.0
alembic>=1.12.1
psutil>=5.9.6
lime>=0.2.0
yellowbrick>=1.5
mlflow>=2.8.0
scikit-optimize>=0.9.0
imbalanced-learn>=0.12.4
feature_engine>=1.6.1
evidently>=0.4.9
alibi-detect>=0.11.3
mlxtend>=0.23.0
umap-learn>=0.5.4
seaborn>=0.13.0
matplotlib>=3.8.2
pillow>=10.1.0
sparse>=0.15.4
skl2onnx>=1.15.3
onnxruntime>=1.14.1
sqlalchemy>=2.0.0
python-multipart>=0.0.9